pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.8",
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-openai>=1.2.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
